[DeLICM] Use opt<int>.
There is no template specialization for cl::parser<unsigned long> such
that parsing an cl::opt<unsigned long> command line argument will fail.
Use opt<int> instead which has an associated parser.

llvm-svn: 295832

+++ polly/lib/Transform/DeLICM.cpp
@@ -140,7 +140,7 @@ using namespace llvm;
 
 namespace {
 
-cl::opt<unsigned long>
+cl::opt<int>
     DelicmMaxOps("polly-delicm-max-ops",
                  cl::desc("Maximum number of isl operations to invest for "
                           "lifetime analysis; 0=no limit"),
